// WorkDiaryWeb/src/app/buyers/models/Job.ts
export interface Job {
  USER_ID: number
  JOB_ID: number
  JOB_TYPE_ID: number
  JOB_TYPE_NAME: string
  JOB_STATUS_ID: number
  WORKING_ON: string
  TEAM_MEMBERS: string
  JOB_STATUS_NAME: string
  JOB_TITLE: string
  DESCRIPTION: string
  APPROXIMATE_DURATION: string
  APPROXIMATE_BUDGET: number
  APPROXIMATE_BUDGET_DISPLAY: string
  DURATION: any
  RATE_TYPE_NAME: string
  CURRENCY_NAME: string
  CURRENCY_ID: number
  SYMBOL: string
  RATE: number
  DISPLAY_RATE: string
  DOLLOR_RATE: string
  HOURLY_RATE: number
  HOURLY_RATE_DISPLAY: string
  HOURS_PER_WEEK: number
  ON_SITE_WORK: boolean
  SITE_LOCATION_COUNTRY_NAME: string
  SITE_LOCATION_REGION_ID: number
  SITE_LOCATION_CITY_ID: number
  SITE_LOCATION_REGION_NAME: string
  SITE_LOCATION_CITY_NAME: string
  SITE_LOCATION_POSTCODE: string
  OFFER_SUBMIT_DAYS: number
  WORK_START: string
  WORK_START_IMMIDIATELY: boolean
  WORK_START_ON_DATE: string
  JOB_VIEW_IS_PUBLIC: boolean
  OFFER_AMOUNT_IS_PUBLIC: boolean
  CREATED_ON: string
  CREATED_ON_DISPLAY: string
  CREATED_BY: number
  IS_ARCHIVED: boolean
  JOB_SKILL_ID: number
  CAT_ID: number
  CATEGORY_NAME: string
  CLASS_ID: number
  CLASS_NAME: string
  SKILL_ID: number
  SKILL_NAME: string
  JOB_SKILLS: string
  BUYER_USER_ID: number
  BUYER_CONTACT_NAME: string
  BUYER_USER_NAME: string
  JOB_OFFER_COUNT: number
  JOB_AWARDED_COUNT: number
  JOB_INVITED_COUNT: number
  MATTERMOST_TEAM_ID: any
  MATTERMOST_TEAM_NAME: any
}